lint fixes, move files around, minor changes
