+++ .env.example
@@ -1,8 +1,6 @@
+GOOGLE_ID=
+GOOGLE_SECRET=
+NEXTAUTH_SECRET=
+NEXTAUTH_URL=
+DATABASE_URL=
 DATABASE_URL=postgres://
-
-
-GOOGLE_ID
-GOOGLE_SECRET
-NEXTAUTH_SECRET
-NEXTAUTH_URL
-DATABASE_URL
