+++ app/login/page.tsx
@@ -3,6 +3,7 @@
 import Image from "next/image";
 import { signIn } from "next-auth/react";
 import AuthButton from "@/components/auth-button";
+
 export default function Login() {
     return (
         <div className="flex w-full flex-col lg:flex-row">
