+++ lib/constants.ts
@@ -2,7 +2,7 @@ export const colorOptions = ["#ED9D9D", "#F3AB7E", "#EEF583", "#94ED79", "#8E87F
 
 export const daysOptions = ["M", "T", "W", "Th", "F"] as const;
 
-export const dayLabels: Record<typeof daysOptions[number], string> = {
+export const dayLabels: Record<(typeof daysOptions)[number], string> = {
     M: "Monday",
     T: "Tuesday",
     W: "Wednesday",
