+++ tsconfig.json
@@ -22,6 +22,6 @@
             "@/*": ["./*"]
         }
     },
-    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "lib/types.d.ts"],
+    "include": ["**/*.ts", "**/*.tsx", "lib/types.d.ts", "next-env.d.ts", ".next/types/**/*.ts"],
     "exclude": ["node_modules", "prisma/generated"]
 }
